backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
celery==5.4.0
redis==5.2.1
psutil==6.1.1
reportlab==4.2.5
